+++ docker-jans-keycloak-link/scripts/upgrade.py
@@ -58,7 +58,7 @@ class Upgrade:
         kwargs = {"table_name": "jansCustomScr"}
         script_id = doc_id_from_dn("inum=13D3-E7AD,ou=scripts,o=jans")
 
-        # toggle cache-refresh script
+        # toggle link-interception script
         entry = self.backend.get_entry(script_id, **kwargs)
 
         if entry: